refactor(Christov): Delete the unoptimized Christov version
