+++ PeakSwift/Sources/PeakSwift/Algorithms/Christov.swift
@@ -2,7 +2,7 @@
 //  File.swift
 //  
 //
-//  Created by Maximilian Kapsecker on 11.01.23.
+//  Created by x on 03.06.23.
 //
 
 import Foundation
@@ -11,8 +11,8 @@ import Accelerate
 /// https://pubmed.ncbi.nlm.nih.gov/15333132/
 class Christov: Algorithm {
     
+    /// Based on https://github.com/neuropsychology/NeuroKit
     func detectPeaks(ecgSignal: [Double], samplingFrequency: Double) -> [UInt] {
-        //TODO optimize and refactor
         
         let c1 = 0.02
         let c2 = 0.028
@@ -24,6 +24,7 @@ class Christov: Algorithm {
         let MA2 = lfilter(ecgSignal: MA1, samplingFrequency: samplingFrequency, c: c2)
         
 
+        // TODO eventually optimize the loop
         var Y: [Double] = []
         for i in 1...(MA2.count-2) {
             Y.append(abs(MA2[i+1] - MA2[i-1]))
@@ -40,40 +41,43 @@ class Christov: Algorithm {
         
         var M = 0.0
         var newM5 = 0.0
-        var M_list: [Double] = []
+        var MList: [Double] = []
         var MM: [Double] = []
         
         let stepInterval = (0.6-1) / Double(ms1200 - ms200 - 1)
-        let M_slope = Array(stride(from: 1, through: 0.6, by: stepInterval))
+        let MSlopes = Array(stride(from: 1, through: 0.6, by: stepInterval))
         
         var F = 0.0
-        var F_List: [Double] = []
+        var FList: [Double] = []
         var R = 0.0
         var RR: [Double] = []
         var Rm = 0
-        var R_list: [Double] = []
+        var RList: [Double] = []
         
         var MFR = 0.0
-        var MFR_list: [Double] = []
+        var MFRList: [Double] = []
         
-        var QRS: [Int] = []
+        // Actually expect rPeak Positions to be UInt. Swift array doesn't support index access with UInt.
+        // Swift array are limited to Int size.
+        // Therefore, for now fallback to int
+        var rPeaks: [Int] = []
         
         
         for i in 0...(MA3.count-1) {
            
             // M
             if Double(i) < (5 * samplingFrequency) {
-                M = 0.6 * maxInRange(array: MA3, from: 0, to: i + 1)
+                M = 0.6 * MathUtils.maxInRange(array: MA3, from: 0, to: i + 1)
                 MM.append(M)
                 if MM.count > 5 {
                     MM.remove(at: 0)
                 }
-            } else if let lastQRS = QRS.last, i < (lastQRS + ms200) {
-                newM5 = 0.6 *  maxInRange(array: MA3, from: lastQRS, to: i)
+            } else if let lastRPeak = rPeaks.last, i < lastRPeak + ms200 {
+                newM5 = 0.6 *  MathUtils.maxInRange(array: MA3, from: lastRPeak, to: i)
                 if let MMLast = MM.last, newM5 > 1.5 * MMLast {
                     newM5 = 1.1 * MMLast
                 }
-            } else if let lastQRS = QRS.last, i == lastQRS + ms200 {
+            } else if let lastRPeak = rPeaks.last, i == lastRPeak + ms200 {
                 if newM5 == 0, let MMLast = MM.last {
                     newM5 = MMLast
                 }
@@ -81,67 +85,57 @@ class Christov: Algorithm {
                 if MM.count > 5 {
                     MM.remove(at: 0)
                 }
-                M = mean(array: MM)
-            } else if let lastQRS = QRS.last, i > lastQRS + ms200, i < lastQRS + ms1200 {
-                M = mean(array: MM) * M_slope[i - (lastQRS + ms200)]
+                M = MathUtils.mean(array: MM)
+            } else if let lastRPeak = rPeaks.last, i > lastRPeak + ms200, i < lastRPeak + ms1200 {
+                M = MathUtils.mean(array: MM) * MSlopes[i - (lastRPeak + ms200)]
             } else {
-                M = 0.6 * mean(array: MM)
+                M = 0.6 * MathUtils.mean(array: MM)
             }
             
             
             // F
             if i > ms350 {
-                let F_Section = Array(MA3[(i - ms350)...i])
-                let max_latest = maxInRange(array: F_Section, from: F_Section.count-ms50, to: F_Section.count)
-                let max_earliest = maxInRange(array: F_Section, from: 0, to: ms50)
-                F += (max_latest - max_earliest) / 150.0
+                let FSection = Array(MA3[(i - ms350)...i])
+                let maxLatest = MathUtils.maxInRange(array: FSection, from: FSection.count-ms50, to: FSection.count)
+                let maxEarliest = MathUtils.maxInRange(array: FSection, from: 0, to: ms50)
+                F += (maxLatest - maxEarliest) / 150.0
             }
             
-            if let QRSLast = QRS.last, i < QRSLast + Int(2.0 / 3.0 * Double(Rm)), i < QRSLast + Rm {
-                let dec = (M - mean(array: MM)) / 1.4
+            if let lastRPeak = rPeaks.last, i < lastRPeak + Int(2.0 / 3.0 * Double(Rm)), i < lastRPeak + Rm {
+                let dec = (M - MathUtils.mean(array: MM)) / 1.4
                 R = dec
             }
             
             MFR = M + F + R
-            M_list.append(M)
-            F_List.append(F)
-            R_list.append(R)
-            MFR_list.append(MFR)
+            MList.append(M)
+            FList.append(F)
+            RList.append(R)
+            MFRList.append(MFR)
             
-            if QRS.isEmpty, MA3[i] > MFR {
-                QRS.append(i)
-            } else if let QRSLast = QRS.last, i > QRSLast + ms200, MA3[i] > MFR {
-                QRS.append(i)
-                if QRS.count > 2 {
-                    RR.append(Double(QRSLast - QRS[QRS.count-2]))
+            if rPeaks.isEmpty, MA3[i] > MFR {
+                rPeaks.append(i)
+            } else if let lastRPeak = rPeaks.last, i > lastRPeak + ms200, MA3[i] > MFR {
+                rPeaks.append(i)
+                if rPeaks.count > 2 {
+                    RR.append(Double(lastRPeak - rPeaks[rPeaks.count-2]))
                     if RR.count > 5 {
                         RR.remove(at: 0)
                     }
-                    Rm = Int(mean(array: RR))
+                    Rm = Int(MathUtils.mean(array: RR))
                 }
             }
             
         }
         
-        QRS.remove(at: 0)
+        rPeaks.remove(at: 0)
 
-        return QRS.map { r in
+        return rPeaks.map { r in
             UInt(r)
         }
     }
     
-    func maxInRange<T>(array: [T], from: Int, to: Int) -> T where T:Comparable & AdditiveArithmetic{
-        let arraySlice = array[from...to-1]
-        let arrayInRange = Array(arraySlice)
-        return arrayInRange.max() ?? T.zero
-    }
     
-    func mean(array: [Double]) -> Double {
-        let sum = array.reduce(0.0, +)
-        return sum / Double(array.count)
-    }
-    
-    func lfilter(ecgSignal: [Double], samplingFrequency: Double, c: Double) -> [Double] {
+    private func lfilter(ecgSignal: [Double], samplingFrequency: Double, c: Double) -> [Double] {
         
         let impulseRespone = createImpulseResponse(samplingFrequency: samplingFrequency, c: c)
         let filteredSignal = FIT().filter(impulseResponse: impulseRespone, signal: ecgSignal)
@@ -149,10 +143,9 @@ class Christov: Algorithm {
 
     }
     
-    func createImpulseResponse(samplingFrequency: Double, c: Double) -> [Double] {
+    private func createImpulseResponse(samplingFrequency: Double, c: Double) -> [Double] {
         let impulseResponeTemp: [Double] = [Double](repeating: 1, count: Int(samplingFrequency * c))
         
-        // TODO optimimize
         let impulseRespone = impulseResponeTemp.map { x in
             x / (samplingFrequency * c)
         }
