+++ PeakSwift/Sources/PeakSwift/Context/AlgorithmStrategy.swift
@@ -30,8 +30,6 @@ class AlgorithmStrategy {
             self.algorithm = WQRS()
         case .Nabian2018:
             self.algorithm = Nabian2018()
-        case .ChristovOptimized:
-            self.algorithm = ChristovOptimized()
         }
     }
